Moves log from lib/run/ to lib/
The logger is needed before the run process begins so it's no longer a
local dep of just the run process.

(For instance, to write errors/warnings during the loading process.)
