+++ lib/derive-log-level.js
@@ -1,4 +1,4 @@
-var level = require('./run/log').levels
+var level = require('./log').levels
 
 module.exports = function deriveLogLevel (userOptions) {
   if (!userOptions) return
