+++ lib/derive-log-level.test.js
@@ -1,5 +1,5 @@
 var subject = require('./derive-log-level')
-var level = require('./run/log').levels
+var level = require('./log').levels
 
 module.exports = {
   'does not default': function () {
