+++ lib/log.test.js
