+++ lib/resolve-script/find-executables.js
@@ -1,6 +1,6 @@
 var _ = require('lodash')
 var globFirst = require('./glob-first')
-var log = require('../run/log')
+var log = require('../log')
 var path = require('path')
 var async = require('async')
