+++ lib/resolve-script/find-executables.test.js
@@ -1,6 +1,6 @@
 var _ = require('lodash')
 var path = require('path')
-var log = require('../run/log')
+var log = require('../log')
 
 var base = function (glob) {
   return path.resolve('test/fixtures/unit/find-executables', glob)
