+++ lib/run/dry-run.js
@@ -1,7 +1,7 @@
 var _ = require('lodash')
 
 var printScript = require('./print-script')
-var log = require('./log')
+var log = require('../log')
 
 module.exports = function (scriptFiles, cb) {
   log.info('This is a dry run. Executed scripts would be:\n')
