+++ lib/run/print-script.js
@@ -1,7 +1,7 @@
 var fs = require('fs')
 var _ = require('lodash')
 
-var log = require('./log')
+var log = require('../log')
 
 module.exports = function (scriptFile) {
   log.info('Executing "' + scriptFile + '":\n')
