+++ lib/run/print-script.test.js
@@ -1,6 +1,6 @@
 var path = require('path')
 
-var log = require('./log')
+var log = require('../log')
 
 module.exports = {
   beforeEach: function () {
