+++ lib/scripty.js
@@ -1,7 +1,7 @@
 var optionify = require('./optionify')
 var resolveScript = require('./resolve-script')
 var run = require('./run')
-var log = require('./run/log')
+var log = require('./log')
 
 module.exports = optionify(function scripty (npmLifecycle, options, cb) {
   log.level(options.logLevel)
