+++ test/safe-helper.js
@@ -1,7 +1,7 @@
 var decorateAssertions = require('./decorate-assertions')
 global.assert = decorateAssertions(require('assert'))
 
-var log = require('../lib/run/log')
+var log = require('../lib/log')
 
 module.exports = {
   afterEach: function () {
