+++ test/safe/basic-test.js
@@ -1,5 +1,5 @@
 var runScripty = require('../run-scripty')
-var log = require('../../lib/run/log')
+var log = require('../../lib/log')
 
 module.exports = {
   outputAndRunScript: function (done) {
