+++ test/unit-helper.js
@@ -3,7 +3,7 @@ global.td = require('testdouble')
 var decorateAssertions = require('./decorate-assertions')
 global.assert = decorateAssertions(require('assert'))
 
-var log = require('../lib/run/log')
+var log = require('../lib/log')
 
 global.UNSUPPORTED_TDD = require('./is-old-node')
 if (UNSUPPORTED_TDD) {
